Refactor: split files
