+++ lib/src/context.rs
@@ -1,0 +1,86 @@
+use crate::surface_wrapper::*;
+use std::sync::Arc;
+use winit::window::Window;
+
+/// Context containing global wgpu resources.
+pub struct Context {
+    pub instance: wgpu::Instance,
+    pub adapter: wgpu::Adapter,
+    pub device: wgpu::Device,
+    pub queue: wgpu::Queue,
+}
+impl Context {
+    /// Initializes the example context.
+    pub async fn init_async(surface: &mut SurfaceWrapper, window: Arc<Window>) -> Self {
+        let backends = wgpu::util::backend_bits_from_env().unwrap_or_default();
+        let dx12_shader_compiler = wgpu::util::dx12_shader_compiler_from_env().unwrap_or_default();
+        let gles_minor_version = wgpu::util::gles_minor_version_from_env().unwrap_or_default();
+
+        let instance = wgpu::Instance::new(wgpu::InstanceDescriptor {
+            backends,
+            flags: wgpu::InstanceFlags::from_build_config().with_env(),
+            dx12_shader_compiler,
+            gles_minor_version,
+        });
+        surface.pre_adapter(&instance, window);
+        let adapter = wgpu::util::initialize_adapter_from_env_or_default(&instance, surface.get())
+            .await
+            .expect("No suitable GPU adapters found on the system!");
+
+        // let adapter_info = adapter.get_info();
+        // log::info!("Using {} ({:?})", adapter_info.name, adapter_info.backend);
+
+        let optional_features = wgpu::Features::empty();
+        let required_features = wgpu::Features::empty();
+        let adapter_features = adapter.features();
+        assert!(
+            adapter_features.contains(required_features),
+            "Adapter does not support required features for this example: {:?}",
+            required_features - adapter_features
+        );
+
+        let required_downlevel_capabilities = wgpu::DownlevelCapabilities {
+            flags: wgpu::DownlevelFlags::empty(),
+            shader_model: wgpu::ShaderModel::Sm5,
+            ..wgpu::DownlevelCapabilities::default()
+        };
+        let downlevel_capabilities = adapter.get_downlevel_capabilities();
+        assert!(
+            downlevel_capabilities.shader_model >= required_downlevel_capabilities.shader_model,
+            "Adapter does not support the minimum shader model required to run this example: {:?}",
+            required_downlevel_capabilities.shader_model
+        );
+        assert!(
+            downlevel_capabilities
+                .flags
+                .contains(required_downlevel_capabilities.flags),
+            "Adapter does not support the downlevel capabilities required to run this example: {:?}",
+            required_downlevel_capabilities.flags - downlevel_capabilities.flags
+        );
+
+        // Make sure we use the texture resolution limits from the adapter, so we can support images the size of the surface.
+        let needed_limits =
+            wgpu::Limits::downlevel_webgl2_defaults().using_resolution(adapter.limits());
+
+        let trace_dir = std::env::var("WGPU_TRACE");
+        let (device, queue) = adapter
+            .request_device(
+                &wgpu::DeviceDescriptor {
+                    label: None,
+                    required_features: (optional_features & adapter_features) | required_features,
+                    required_limits: needed_limits,
+                    ..Default::default()
+                },
+                trace_dir.ok().as_ref().map(std::path::Path::new),
+            )
+            .await
+            .expect("Unable to find a suitable GPU adapter!");
+
+        Self {
+            instance,
+            adapter,
+            device,
+            queue,
+        }
+    }
+}
