+++ lib/src/lib.rs
@@ -1,12 +1,6 @@
-use bytemuck::{Pod, Zeroable};
-use std::{
-    borrow::Cow, cell::RefCell, f32::consts, marker::PhantomData, mem, num::NonZeroU8, rc::Rc,
-    sync::Arc,
-};
-use wgpu::{util::DeviceExt, Instance, Surface};
+use std::{cell::RefCell, marker::PhantomData, num::NonZeroU8, rc::Rc, sync::Arc};
 use winit::{
-    dpi::PhysicalSize,
-    event::{ElementState, Event, KeyEvent, StartCause, WindowEvent},
+    event::{Event, KeyEvent, WindowEvent},
     event_loop::{EventLoop, EventLoopWindowTarget},
     keyboard::{Key, NamedKey},
     window::Window,
@@ -26,11 +20,17 @@ use core_graphics::{
 #[cfg(target_arch = "wasm32")]
 use wasm_bindgen::prelude::*;
 
+mod context;
 mod input;
 mod map;
+mod surface_wrapper;
+mod vertex;
+mod vox;
 
+use context::Context;
 use input::Input;
-use map::{Cube, Map, CHUNK_SIZE};
+use surface_wrapper::SurfaceWrapper;
+use vox::*;
 
 struct EventLoopWrapper {
     event_loop: EventLoop<()>,
@@ -84,233 +84,6 @@ impl EventLoopWrapper {
             event_loop,
             window,
             canvas: outer_canvas,
-        }
-    }
-}
-
-/// Wrapper type which manages the surface and surface configuration.
-///
-/// As surface usage varies per platform, wrapping this up cleans up the event loop code.
-struct SurfaceWrapper {
-    surface: Option<wgpu::Surface<'static>>,
-    config: Option<wgpu::SurfaceConfiguration>,
-}
-
-impl SurfaceWrapper {
-    /// Create a new surface wrapper with no surface or configuration.
-    fn new() -> Self {
-        Self {
-            surface: None,
-            config: None,
-        }
-    }
-
-    /// Called after the instance is created, but before we request an adapter.
-    ///
-    /// On wasm, we need to create the surface here, as the WebGL backend needs
-    /// a surface (and hence a canvas) to be present to create the adapter.
-    ///
-    /// We cannot unconditionally create a surface here, as Android requires
-    /// us to wait until we receive the `Resumed` event to do so.
-    fn pre_adapter(&mut self, instance: &Instance, window: Arc<Window>) {
-        if cfg!(target_arch = "wasm32") {
-            self.surface = Some(instance.create_surface(window).unwrap());
-        }
-    }
-
-    /// Check if the event is the start condition for the surface.
-    fn start_condition(e: &Event<()>) -> bool {
-        match e {
-            // On all other platforms, we can create the surface immediately.
-            Event::NewEvents(StartCause::Init) => !cfg!(target_os = "android"),
-            // On android we need to wait for a resumed event to create the surface.
-            Event::Resumed => cfg!(target_os = "android"),
-            _ => false,
-        }
-    }
-
-    /// Called when an event which matches [`Self::start_condition`] is received.
-    ///
-    /// On all native platforms, this is where we create the surface.
-    ///
-    /// Additionally, we configure the surface based on the (now valid) window size.
-    fn resume(&mut self, context: &Context, window: Arc<Window>, srgb: bool) {
-        // Window size is only actually valid after we enter the event loop.
-        let window_size = window.inner_size();
-        let width = window_size.width.max(1);
-        let height = window_size.height.max(1);
-
-        log::info!("Surface resume {window_size:?}");
-
-        // We didn't create the surface in pre_adapter, so we need to do so now.
-        if !cfg!(target_arch = "wasm32") {
-            self.surface = Some(context.instance.create_surface(window).unwrap());
-        }
-
-        // From here on, self.surface should be Some.
-
-        let surface = self.surface.as_ref().unwrap();
-
-        // Get the default configuration,
-        let mut config = surface
-            .get_default_config(&context.adapter, width, height)
-            .expect("Surface isn't supported by the adapter.");
-        if srgb {
-            // Not all platforms (WebGPU) support sRGB swapchains, so we need to use view formats
-            let view_format = config.format.add_srgb_suffix();
-            config.view_formats.push(view_format);
-        } else {
-            // All platforms support non-sRGB swapchains, so we can just use the format directly.
-            let format = config.format.remove_srgb_suffix();
-            config.format = format;
-            config.view_formats.push(format);
-        };
-
-        surface.configure(&context.device, &config);
-        self.config = Some(config);
-    }
-
-    /// Resize the surface, making sure to not resize to zero.
-    fn resize(&mut self, context: &Context, size: PhysicalSize<u32>) {
-        log::info!("Surface resize {size:?}");
-
-        let config = self.config.as_mut().unwrap();
-        config.width = size.width;
-        config.height = size.height;
-        #[cfg(target_arch = "wasm32")]
-        {
-            let device_pixel_ratio = web_sys::window().unwrap().device_pixel_ratio();
-            config.width = (config.width as f64 / device_pixel_ratio) as u32;
-            config.height = (config.height as f64 / device_pixel_ratio) as u32;
-        }
-        config.width = config.width.max(1);
-        config.height = config.height.max(1);
-        let surface = self.surface.as_ref().unwrap();
-        surface.configure(&context.device, config);
-    }
-
-    /// Acquire the next surface texture.
-    fn acquire(&mut self, context: &Context) -> wgpu::SurfaceTexture {
-        let surface = self.surface.as_ref().unwrap();
-
-        match surface.get_current_texture() {
-            Ok(frame) => frame,
-            // If we timed out, just try again
-            Err(wgpu::SurfaceError::Timeout) => surface
-                .get_current_texture()
-                .expect("Failed to acquire next surface texture!"),
-            Err(
-                // If the surface is outdated, or was lost, reconfigure it.
-                wgpu::SurfaceError::Outdated
-                | wgpu::SurfaceError::Lost
-                // If OutOfMemory happens, reconfiguring may not help, but we might as well try
-                | wgpu::SurfaceError::OutOfMemory,
-            ) => {
-                surface.configure(&context.device, self.config());
-                surface
-                    .get_current_texture()
-                    .expect("Failed to acquire next surface texture!")
-            }
-        }
-    }
-
-    /// On suspend on android, we drop the surface, as it's no longer valid.
-    ///
-    /// A suspend event is always followed by at least one resume event.
-    fn suspend(&mut self) {
-        if cfg!(target_os = "android") {
-            self.surface = None;
-        }
-    }
-
-    fn get(&self) -> Option<&Surface> {
-        self.surface.as_ref()
-    }
-
-    fn config(&self) -> &wgpu::SurfaceConfiguration {
-        self.config.as_ref().unwrap()
-    }
-}
-
-/// Context containing global wgpu resources.
-struct Context {
-    instance: wgpu::Instance,
-    adapter: wgpu::Adapter,
-    device: wgpu::Device,
-    queue: wgpu::Queue,
-}
-impl Context {
-    /// Initializes the example context.
-    async fn init_async(surface: &mut SurfaceWrapper, window: Arc<Window>) -> Self {
-        let backends = wgpu::util::backend_bits_from_env().unwrap_or_default();
-        let dx12_shader_compiler = wgpu::util::dx12_shader_compiler_from_env().unwrap_or_default();
-        let gles_minor_version = wgpu::util::gles_minor_version_from_env().unwrap_or_default();
-
-        let instance = wgpu::Instance::new(wgpu::InstanceDescriptor {
-            backends,
-            flags: wgpu::InstanceFlags::from_build_config().with_env(),
-            dx12_shader_compiler,
-            gles_minor_version,
-        });
-        surface.pre_adapter(&instance, window);
-        let adapter = wgpu::util::initialize_adapter_from_env_or_default(&instance, surface.get())
-            .await
-            .expect("No suitable GPU adapters found on the system!");
-
-        // let adapter_info = adapter.get_info();
-        // log::info!("Using {} ({:?})", adapter_info.name, adapter_info.backend);
-
-        let optional_features = wgpu::Features::empty();
-        let required_features = wgpu::Features::empty();
-        let adapter_features = adapter.features();
-        assert!(
-            adapter_features.contains(required_features),
-            "Adapter does not support required features for this example: {:?}",
-            required_features - adapter_features
-        );
-
-        let required_downlevel_capabilities = wgpu::DownlevelCapabilities {
-            flags: wgpu::DownlevelFlags::empty(),
-            shader_model: wgpu::ShaderModel::Sm5,
-            ..wgpu::DownlevelCapabilities::default()
-        };
-        let downlevel_capabilities = adapter.get_downlevel_capabilities();
-        assert!(
-            downlevel_capabilities.shader_model >= required_downlevel_capabilities.shader_model,
-            "Adapter does not support the minimum shader model required to run this example: {:?}",
-            required_downlevel_capabilities.shader_model
-        );
-        assert!(
-            downlevel_capabilities
-                .flags
-                .contains(required_downlevel_capabilities.flags),
-            "Adapter does not support the downlevel capabilities required to run this example: {:?}",
-            required_downlevel_capabilities.flags - downlevel_capabilities.flags
-        );
-
-        // Make sure we use the texture resolution limits from the adapter, so we can support images the size of the surface.
-        let needed_limits =
-            wgpu::Limits::downlevel_webgl2_defaults().using_resolution(adapter.limits());
-
-        let trace_dir = std::env::var("WGPU_TRACE");
-        let (device, queue) = adapter
-            .request_device(
-                &wgpu::DeviceDescriptor {
-                    label: None,
-                    required_features: (optional_features & adapter_features) | required_features,
-                    required_limits: needed_limits,
-                    ..Default::default()
-                },
-                trace_dir.ok().as_ref().map(std::path::Path::new),
-            )
-            .await
-            .expect("Unable to find a suitable GPU adapter!");
-
-        Self {
-            instance,
-            adapter,
-            device,
-            queue,
         }
     }
 }
@@ -557,450 +330,4 @@ pub async fn run() {
             }
         },
     );
-}
-
-#[repr(C)]
-#[derive(Clone, Copy, Pod, Zeroable)]
-struct Vertex {
-    _pos: [f32; 4],
-    _tex_coord: [f32; 2],
-}
-
-fn vertex(pos: [f32; 3], tc: [f32; 2]) -> Vertex {
-    Vertex {
-        _pos: [pos[0], pos[1], pos[2], 1.0],
-        _tex_coord: [tc[0], tc[1]],
-    }
-}
-
-fn create_vertices(x: f32, y: f32, z: f32, index: usize) -> (Vec<Vertex>, Vec<u16>) {
-    let vertex_data = [
-        // top (0, 0, 1)
-        vertex([x + 0.0, y + 0.0, z + 1.0], [0.0, 0.0]),
-        vertex([x + 1.0, y + 0.0, z + 1.0], [1.0, 0.0]),
-        vertex([x + 1.0, y + 1.0, z + 1.0], [1.0, 1.0]),
-        vertex([x + 0.0, y + 1.0, z + 1.0], [0.0, 1.0]),
-        // bottom (0, 0, -1)
-        vertex([x + 0.0, y + 1.0, z + 0.0], [1.0, 0.0]),
-        vertex([x + 1.0, y + 1.0, z + 0.0], [0.0, 0.0]),
-        vertex([x + 1.0, y + 0.0, z + 0.0], [0.0, 1.0]),
-        vertex([x + 0.0, y + 0.0, z + 0.0], [1.0, 1.0]),
-        // right (1, 0, 0)
-        vertex([x + 1.0, y + 0.0, z + 0.0], [0.0, 0.0]),
-        vertex([x + 1.0, y + 1.0, z + 0.0], [1.0, 0.0]),
-        vertex([x + 1.0, y + 1.0, z + 1.0], [1.0, 1.0]),
-        vertex([x + 1.0, y + 0.0, z + 1.0], [0.0, 1.0]),
-        // left (-1, 0, 0)
-        vertex([x + 0.0, y + 0.0, z + 1.0], [1.0, 0.0]),
-        vertex([x + 0.0, y + 1.0, z + 1.0], [0.0, 0.0]),
-        vertex([x + 0.0, y + 1.0, z + 0.0], [0.0, 1.0]),
-        vertex([x + 0.0, y + 0.0, z + 0.0], [1.0, 1.0]),
-        // front (0, 1, 0)
-        vertex([x + 1.0, y + 1.0, z + 0.0], [1.0, 0.0]),
-        vertex([x + 0.0, y + 1.0, z + 0.0], [0.0, 0.0]),
-        vertex([x + 0.0, y + 1.0, z + 1.0], [0.0, 1.0]),
-        vertex([x + 1.0, y + 1.0, z + 1.0], [1.0, 1.0]),
-        // back (0, -1, 0)
-        vertex([x + 1.0, y + 0.0, z + 1.0], [0.0, 0.0]),
-        vertex([x + 0.0, y + 0.0, z + 1.0], [1.0, 0.0]),
-        vertex([x + 0.0, y + 0.0, z + 0.0], [1.0, 1.0]),
-        vertex([x + 1.0, y + 0.0, z + 0.0], [0.0, 1.0]),
-    ];
-
-    let offset = index as u16 * 24;
-    let index_data: &[u16] = &[
-        offset,
-        1 + offset,
-        2 + offset,
-        2 + offset,
-        3 + offset,
-        offset, // top
-        4 + offset,
-        5 + offset,
-        6 + offset,
-        6 + offset,
-        7 + offset,
-        4 + offset, // bottom
-        8 + offset,
-        9 + offset,
-        10 + offset,
-        10 + offset,
-        11 + offset,
-        8 + offset, // right
-        12 + offset,
-        13 + offset,
-        14 + offset,
-        14 + offset,
-        15 + offset,
-        12 + offset, // left
-        16 + offset,
-        17 + offset,
-        18 + offset,
-        18 + offset,
-        19 + offset,
-        16 + offset, // front
-        20 + offset,
-        21 + offset,
-        22 + offset,
-        22 + offset,
-        23 + offset,
-        20 + offset, // back
-    ];
-
-    (vertex_data.to_vec(), index_data.to_vec())
-}
-
-fn create_texels(size: usize) -> Vec<u8> {
-    (0..size * size)
-        .map(|id| {
-            // get high five for recognizing this ;)
-            let cx = 3.0 * (id % size) as f32 / (size - 1) as f32 - 2.0;
-            let cy = 2.0 * (id / size) as f32 / (size - 1) as f32 - 1.0;
-            let (mut x, mut y, mut count) = (cx, cy, 0);
-            while count < 0xFF && x * x + y * y < 4.0 {
-                let old_x = x;
-                x = x * x - y * y + cx;
-                y = 2.0 * old_x * y + cy;
-                count += 1;
-            }
-            count
-        })
-        .collect()
-}
-
-#[repr(C)]
-#[derive(Copy, Clone, Pod, Zeroable)]
-struct Uniforms {
-    transform: [[f32; 4]; 4], // This represents a single transformation matrix.
-}
-
-struct Vox {
-    eye: glam::Vec3,
-    horizontal_rotation: f32,
-    vertical_rotation: f32,
-    projection_matrix: glam::Mat4,
-    depth_buffer: wgpu::TextureView,
-    vertex_buf: wgpu::Buffer,
-    index_buf: wgpu::Buffer,
-    index_count: u32,
-    bind_group: wgpu::BindGroup,
-    uniform_vp_buffer: wgpu::Buffer,
-    pipeline: wgpu::RenderPipeline,
-}
-
-impl Vox {
-    fn generate_projection_matrix(aspect_ratio: f32) -> glam::Mat4 {
-        glam::Mat4::perspective_rh(consts::FRAC_PI_4, aspect_ratio, 1.0, 1000.0)
-    }
-
-    fn init(
-        config: &wgpu::SurfaceConfiguration,
-        _adapter: &wgpu::Adapter,
-        device: &wgpu::Device,
-        queue: &wgpu::Queue,
-    ) -> Self {
-        let texture_extent = wgpu::Extent3d {
-            width: config.width,
-            height: config.height,
-            depth_or_array_layers: 1,
-        };
-        let draw_depth_buffer = device.create_texture(&wgpu::TextureDescriptor {
-            label: Some("Depth Buffer"),
-            size: texture_extent,
-            mip_level_count: 1,
-            sample_count: 1,
-            dimension: wgpu::TextureDimension::D2,
-            format: wgpu::TextureFormat::Depth32Float,
-            usage: wgpu::TextureUsages::TEXTURE_BINDING
-                | wgpu::TextureUsages::COPY_DST
-                | wgpu::TextureUsages::RENDER_ATTACHMENT,
-            view_formats: &[],
-        });
-        let depth_buffer = draw_depth_buffer.create_view(&wgpu::TextureViewDescriptor::default());
-
-        let chunk: map::Chunk = Map::new(42).get_chunk(0, 0, 0);
-        let mut cubes = Vec::new();
-        for z in 0..CHUNK_SIZE {
-            for y in 0..CHUNK_SIZE {
-                for x in 0..CHUNK_SIZE {
-                    if chunk.cubes[z * CHUNK_SIZE * CHUNK_SIZE + y * CHUNK_SIZE + x]
-                        == Cube::OnlyOneAtThisTime
-                    {
-                        cubes.push([x as f32, y as f32, z as f32]);
-                    }
-                }
-            }
-        }
-
-        let mut vertex_data = Vec::<Vertex>::new();
-        let mut index_data = Vec::<u16>::new();
-        for (i, &[x, y, z]) in cubes.iter().enumerate() {
-            let (mut vertex, mut index) = create_vertices(x, y, z, i);
-            vertex_data.append(&mut vertex);
-            index_data.append(&mut index);
-        }
-
-        // Create the vertex and index buffers
-        let vertex_size = mem::size_of::<Vertex>();
-
-        let vertex_buf = device.create_buffer_init(&wgpu::util::BufferInitDescriptor {
-            label: Some("Vertex Buffer"),
-            contents: bytemuck::cast_slice(&vertex_data),
-            usage: wgpu::BufferUsages::VERTEX,
-        });
-
-        let index_buf = device.create_buffer_init(&wgpu::util::BufferInitDescriptor {
-            label: Some("Index Buffer"),
-            contents: bytemuck::cast_slice(&index_data),
-            usage: wgpu::BufferUsages::INDEX,
-        });
-
-        // Create pipeline layout
-        let bind_group_layout = device.create_bind_group_layout(&wgpu::BindGroupLayoutDescriptor {
-            label: None,
-            entries: &[
-                wgpu::BindGroupLayoutEntry {
-                    binding: 0,
-                    visibility: wgpu::ShaderStages::VERTEX,
-                    ty: wgpu::BindingType::Buffer {
-                        ty: wgpu::BufferBindingType::Uniform,
-                        has_dynamic_offset: false,
-                        min_binding_size: wgpu::BufferSize::new(64),
-                    },
-                    count: None,
-                },
-                wgpu::BindGroupLayoutEntry {
-                    binding: 1,
-                    visibility: wgpu::ShaderStages::FRAGMENT,
-                    ty: wgpu::BindingType::Texture {
-                        multisampled: false,
-                        sample_type: wgpu::TextureSampleType::Uint,
-                        view_dimension: wgpu::TextureViewDimension::D2,
-                    },
-                    count: None,
-                },
-            ],
-        });
-        let pipeline_layout = device.create_pipeline_layout(&wgpu::PipelineLayoutDescriptor {
-            label: None,
-            bind_group_layouts: &[&bind_group_layout],
-            push_constant_ranges: &[],
-        });
-
-        // Create the texture
-        let size = 256u32;
-        let texels = create_texels(size as usize);
-        let texture_extent = wgpu::Extent3d {
-            width: size,
-            height: size,
-            depth_or_array_layers: 1,
-        };
-        let texture = device.create_texture(&wgpu::TextureDescriptor {
-            label: None,
-            size: texture_extent,
-            mip_level_count: 1,
-            sample_count: 1,
-            dimension: wgpu::TextureDimension::D2,
-            format: wgpu::TextureFormat::R8Uint,
-            usage: wgpu::TextureUsages::TEXTURE_BINDING | wgpu::TextureUsages::COPY_DST,
-            view_formats: &[],
-        });
-        let texture_view = texture.create_view(&wgpu::TextureViewDescriptor::default());
-        queue.write_texture(
-            texture.as_image_copy(),
-            &texels,
-            wgpu::ImageDataLayout {
-                offset: 0,
-                bytes_per_row: Some(size),
-                rows_per_image: None,
-            },
-            texture_extent,
-        );
-
-        let min_alignment = device.limits().min_uniform_buffer_offset_alignment as usize;
-        let uniform_size = std::mem::size_of::<Uniforms>();
-        let aligned_uniform_size =
-            ((uniform_size + min_alignment - 1) / min_alignment) * min_alignment;
-
-        // Create other resources
-        let uniform_vp_buffer = device.create_buffer(&wgpu::BufferDescriptor {
-            label: Some("Uniform VP Buffer"),
-            usage: wgpu::BufferUsages::UNIFORM | wgpu::BufferUsages::COPY_DST,
-            size: aligned_uniform_size as wgpu::BufferAddress,
-            mapped_at_creation: false,
-        });
-
-        let bind_group = device.create_bind_group(&wgpu::BindGroupDescriptor {
-            layout: &bind_group_layout,
-            entries: &[
-                wgpu::BindGroupEntry {
-                    binding: 0,
-                    resource: uniform_vp_buffer.as_entire_binding(),
-                },
-                wgpu::BindGroupEntry {
-                    binding: 1,
-                    resource: wgpu::BindingResource::TextureView(&texture_view),
-                },
-            ],
-            label: None,
-        });
-
-        let shader = device.create_shader_module(wgpu::ShaderModuleDescriptor {
-            label: None,
-            source: wgpu::ShaderSource::Wgsl(Cow::Borrowed(include_str!("shader.wgsl"))),
-        });
-
-        let vertex_buffers = [wgpu::VertexBufferLayout {
-            array_stride: vertex_size as wgpu::BufferAddress,
-            step_mode: wgpu::VertexStepMode::Vertex,
-            attributes: &[
-                wgpu::VertexAttribute {
-                    format: wgpu::VertexFormat::Float32x4,
-                    offset: 0,
-                    shader_location: 0,
-                },
-                wgpu::VertexAttribute {
-                    format: wgpu::VertexFormat::Float32x2,
-                    offset: 4 * 4,
-                    shader_location: 1,
-                },
-            ],
-        }];
-
-        let pipeline = device.create_render_pipeline(&wgpu::RenderPipelineDescriptor {
-            label: None,
-            layout: Some(&pipeline_layout),
-            vertex: wgpu::VertexState {
-                module: &shader,
-                entry_point: "vs_main",
-                buffers: &vertex_buffers,
-                compilation_options: wgpu::PipelineCompilationOptions::default(),
-            },
-            fragment: Some(wgpu::FragmentState {
-                module: &shader,
-                entry_point: "fs_main",
-                targets: &[Some(config.view_formats[0].into())],
-                compilation_options: wgpu::PipelineCompilationOptions::default(),
-            }),
-            primitive: wgpu::PrimitiveState {
-                cull_mode: Some(wgpu::Face::Back),
-                ..Default::default()
-            },
-            depth_stencil: Some(wgpu::DepthStencilState {
-                format: wgpu::TextureFormat::Depth32Float,
-                depth_write_enabled: true,
-                depth_compare: wgpu::CompareFunction::Less,
-                stencil: wgpu::StencilState::default(),
-                bias: wgpu::DepthBiasState::default(),
-            }),
-            multisample: wgpu::MultisampleState::default(),
-            multiview: None,
-            cache: None,
-        });
-
-        // Done
-        Vox {
-            eye: glam::Vec3::new(0.0, -5.0, 3.0),
-            horizontal_rotation: 0.0,
-            vertical_rotation: 0.0,
-            projection_matrix: Self::generate_projection_matrix(
-                config.width as f32 / config.height as f32,
-            ),
-            depth_buffer,
-            vertex_buf,
-            index_buf,
-            index_count: index_data.len() as u32,
-            bind_group,
-            uniform_vp_buffer,
-            pipeline,
-        }
-    }
-
-    fn update(&mut self, _event: winit::event::WindowEvent) {
-        //empty
-    }
-
-    fn resize(
-        &mut self,
-        config: &wgpu::SurfaceConfiguration,
-        device: &wgpu::Device,
-        _queue: &wgpu::Queue,
-    ) {
-        let texture_extent = wgpu::Extent3d {
-            width: config.width,
-            height: config.height,
-            depth_or_array_layers: 1,
-        };
-        let draw_depth_buffer = device.create_texture(&wgpu::TextureDescriptor {
-            label: Some("Depth Buffer"),
-            size: texture_extent,
-            mip_level_count: 1,
-            sample_count: 1,
-            dimension: wgpu::TextureDimension::D2,
-            format: wgpu::TextureFormat::Depth32Float,
-            usage: wgpu::TextureUsages::TEXTURE_BINDING
-                | wgpu::TextureUsages::COPY_DST
-                | wgpu::TextureUsages::RENDER_ATTACHMENT,
-            view_formats: &[],
-        });
-        let depth_buffer = draw_depth_buffer.create_view(&wgpu::TextureViewDescriptor::default());
-        self.depth_buffer = depth_buffer;
-
-        self.projection_matrix =
-            Self::generate_projection_matrix(config.width as f32 / config.height as f32);
-    }
-
-    fn render(&mut self, view: &wgpu::TextureView, device: &wgpu::Device, queue: &wgpu::Queue) {
-        let mut encoder =
-            device.create_command_encoder(&wgpu::CommandEncoderDescriptor { label: None });
-
-        {
-            let dir = (glam::Mat3::from_rotation_z(self.horizontal_rotation)
-                * glam::Mat3::from_rotation_x(self.vertical_rotation))
-                * glam::Vec3::Y;
-            let view_matrix = glam::Mat4::look_to_rh(self.eye, dir, glam::Vec3::Z);
-            let mx_total = self.projection_matrix * view_matrix;
-            let mx_ref: &[f32; 16] = mx_total.as_ref();
-            queue.write_buffer(&self.uniform_vp_buffer, 0, bytemuck::cast_slice(mx_ref));
-        }
-
-        {
-            let mut rpass = encoder.begin_render_pass(&wgpu::RenderPassDescriptor {
-                label: None,
-                color_attachments: &[Some(wgpu::RenderPassColorAttachment {
-                    view,
-                    resolve_target: None,
-                    ops: wgpu::Operations {
-                        load: wgpu::LoadOp::Clear(wgpu::Color {
-                            r: 0.1,
-                            g: 0.2,
-                            b: 0.3,
-                            a: 1.0,
-                        }),
-                        store: wgpu::StoreOp::Store,
-                    },
-                })],
-                depth_stencil_attachment: Some(wgpu::RenderPassDepthStencilAttachment {
-                    view: &self.depth_buffer,
-                    depth_ops: Some(wgpu::Operations {
-                        load: wgpu::LoadOp::Clear(1.0),
-                        store: wgpu::StoreOp::Store,
-                    }),
-                    stencil_ops: None,
-                }),
-                timestamp_writes: None,
-                occlusion_query_set: None,
-            });
-            rpass.push_debug_group("Prepare data for draw.");
-            rpass.set_pipeline(&self.pipeline);
-            rpass.set_index_buffer(self.index_buf.slice(..), wgpu::IndexFormat::Uint16);
-            rpass.set_vertex_buffer(0, self.vertex_buf.slice(..));
-            rpass.pop_debug_group();
-            rpass.insert_debug_marker("Draw!");
-            rpass.set_bind_group(0, &self.bind_group, &[]);
-            rpass.draw_indexed(0..self.index_count, 0, 0..1);
-        }
-
-        queue.submit(Some(encoder.finish()));
-    }
 }
